dynamic dashboard path
DASHBOARD_PATH = "/dashboard/"
