+++ .env.example
@@ -11,6 +11,7 @@ UVICORN_PORT = 8000
 # UVICORN_SSL_CERTFILE = "/var/lib/marzban/certs/example.com/fullchain.pem"
 # UVICORN_SSL_KEYFILE = "/var/lib/marzban/certs/example.com/key.pem"
 
+# DASHBOARD_PATH = "/dashboard/"
 
 # XRAY_JSON = "xray_config.json"
 # XRAY_SUBSCRIPTION_URL_PREFIX = "https://example.com"
