+++ .github/workflows/build-dev.yml
@@ -23,7 +23,7 @@ jobs:
 
       - name: Build project
         working-directory: ./app/dashboard
-        run: VITE_BASE_API=/api/ npm run build --if-present -- --outDir build --base '/dashboard/'
+        run: VITE_BASE_API=/api/ npm run build --if-present -- --outDir build --assetsDir statics
 
       - name: Create 404.html
         working-directory: ./app/dashboard
