+++ app/dashboard/__init__.py
@@ -4,17 +4,17 @@ import subprocess
 from pathlib import Path
 
 from app import app
-from config import DEBUG, VITE_BASE_API
+from config import DEBUG, VITE_BASE_API, DASHBOARD_PATH
 from fastapi.staticfiles import StaticFiles
 
-path = '/dashboard/'
 base_dir = Path(__file__).parent
 build_dir = base_dir / 'build'
+statics_dir = build_dir / 'statics'
 
 
 def build():
     proc = subprocess.Popen(
-        ['npm', 'run', 'build', '--', '--base', path,  '--outDir', build_dir],
+        ['npm', 'run', 'build', '--',  '--outDir', build_dir, '--assetsDir', 'statics'],
         env={**os.environ, 'VITE_BASE_API': VITE_BASE_API},
         cwd=base_dir
     )
@@ -27,7 +27,7 @@ def build():
 
 def run_dev():
     proc = subprocess.Popen(
-        ['npm', 'run', 'dev', '--', '--host', '0.0.0.0', '--base', path, '--clearScreen', 'false'],
+        ['npm', 'run', 'dev', '--', '--host', '0.0.0.0', '--clearScreen', 'false'],
         env={**os.environ, 'VITE_BASE_API': VITE_BASE_API},
         cwd=base_dir
     )
@@ -40,9 +40,14 @@ def run_build():
         build()
 
     app.mount(
-        path,
+        DASHBOARD_PATH,
         StaticFiles(directory=build_dir, html=True),
         name="dashboard"
+    )
+    app.mount(
+        '/statics/',
+        StaticFiles(directory=statics_dir, html=True),
+        name="statics"
     )
