+++ app/dashboard/index.html
@@ -3,36 +3,35 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta name="viewport" content="width=device-width, user-scalable=no" />
     <title>Marzban</title>
     <link
       rel="apple-touch-icon"
       sizes="180x180"
-      href="/favicon/apple-touch-icon.png"
+      href="/statics/favicon/apple-touch-icon.png"
     />
     <link
       rel="icon"
       type="image/png"
       sizes="32x32"
-      href="/favicon/favicon-32x32.png"
+      href="/statics/favicon/favicon-32x32.png"
     />
     <link
       rel="icon"
       type="image/png"
       sizes="16x16"
-      href="/favicon/favicon-16x16.png"
+      href="/statics/favicon/favicon-16x16.png"
     />
-    <link rel="manifest" href="/favicon/site.webmanifest" />
+    <link rel="manifest" href="/statics/favicon/site.webmanifest" />
     <link
       rel="mask-icon"
-      href="/favicon/safari-pinned-tab.svg"
+      href="/statics/favicon/safari-pinned-tab.svg"
       color="#5bbad5"
     />
-    <link rel="shortcut icon" href="/favicon/favicon.ico" />
+    <link rel="shortcut icon" href="/statics/favicon/favicon.ico" />
     <meta name="apple-mobile-web-app-title" content="Marzban" />
     <meta name="application-name" content="Marzban" />
     <meta name="msapplication-TileColor" content="#2b5797" />
-    <meta name="msapplication-config" content="/favicon/browserconfig.xml" />
+    <meta name="msapplication-config" content="/statics/favicon/browserconfig.xml" />
     <meta name="theme-color" content="#3B81F6" />
   </head>
   <body>
