+++ app/dashboard/public/statics/favicon/apple-touch-icon.png
