+++ app/dashboard/public/statics/favicon/browserconfig.xml
