+++ app/dashboard/public/statics/favicon/favicon-32x32.png
