+++ app/dashboard/public/statics/favicon/favicon.ico
