+++ app/dashboard/public/statics/favicon/site.webmanifest
