+++ app/dashboard/public/statics/locales/en.json
@@ -50,6 +50,10 @@
   "hostsDialog.fragment.info": "Correct pattern: length,interval,packets",
   "hostsDialog.fragment.info.attention": "Attention: currently, this feature only supported in streisand >= 1.6.12 and v2rayNG >= 1.8.16 (custom config)",
   "hostsDialog.fragment.info.examples": "Examples:",
+  "hostsDialog.noise": "Noise pattern",
+  "hostsDialog.noise.info": "Correct pattern: packets,delay",
+  "hostsDialog.noise.info.attention": "Attention: currently, this feature only supported in streisand >= 1.6.32 and v2rayNG >= 1.8.39 (custom config)",
+  "hostsDialog.noise.info.examples": "Examples:",
   "hostsDialog.host": "Request Host",
   "hostsDialog.host.info": "By default, if a request host is set in the Xray config, this host is used. However, you can set a custom request host here if needed.",
   "hostsDialog.host.multiHost": "To set multiple addresses, separate them with <badge>,</badge> Each time an address is chosen randomly.",
