+++ app/dashboard/public/statics/locales/ru.json
@@ -50,6 +50,10 @@
   "hostsDialog.fragment.info": "length,interval,packet (e.g. 10-100,100-200,tlshello)",
   "hostsDialog.fragment.info.attention": "Attention: currently, this feature only supported in streisand >= 1.6.12 and v2rayNG >= 1.8.16 (custom config)",
   "hostsDialog.fragment.info.examples": "Examples:",
+  "hostsDialog.noise": "Шумовой паттерн",
+  "hostsDialog.noise.info": "packet,delay (e.g. rand:10-20,100-200)",
+  "hostsDialog.noise.info.attention": "Attention: currently, this feature only supported in streisand >= 1.6.32 and v2rayNG >= 1.8.39 (custom config)",
+  "hostsDialog.noise.info.examples": "Examples:",
   "hostsDialog.host": "Host",
   "hostsDialog.host.info": "По умолчанию, если в конфигурации XRAY задан запрашиваемый хост, то он и будет использоваться. Однако, если необходимо, вы можете установить здесь пользовательский запрашиваемый хост.",
   "hostsDialog.host.multiHost": "Чтобы установить несколько адресов, разделяйте их с помощью <badge>,</badge>. Каждый раз будет выбран случайный адрес.",
