+++ app/dashboard/public/statics/locales/zh.json
@@ -1,20 +1,20 @@
 {
-  "active": "active",
+  "active": "激活",
   "activeUsers": "活跃用户",
   "apply": "应用",
   "cancel": "取消",
-  "core.configuration": "Configuration",
-  "core.generalErrorMessage": "Something went wrong, please check the configuration",
-  "core.logs": "Logs",
-  "core.restartCore": "Restart Core",
-  "core.restarting": "Restarting...",
-  "core.save": "Save",
-  "core.socket.closed": "Closed",
-  "core.socket.connected": "Connected",
-  "core.socket.connecting": "Connecting...",
-  "core.socket.not_connected": "Not Connected",
-  "core.successMessage": "Core settings updated successfully",
-  "core.title": "Core Settings",
+  "core.configuration": "配置",
+  "core.generalErrorMessage": "配置有误, 请检查",
+  "core.logs": "日志",
+  "core.restartCore": "重启核心",
+  "core.restarting": "重启中...",
+  "core.save": "保存",
+  "core.socket.closed": "已关闭",
+  "core.socket.connected": "已连接",
+  "core.socket.connecting": "连接中...",
+  "core.socket.not_connected": "未连接",
+  "core.successMessage": "核心设置更新成功",
+  "core.title": "核心设置",
   "createNewUser": "创建新用户",
   "createUser": "创建用户",
   "dataUsage": "总流量",
@@ -26,7 +26,7 @@
   "deleteUser.deleteSuccess": "{{username}} 删除成功。",
   "deleteUser.prompt": "您确定你要删除 <b>{{username}}</b>？",
   "deleteUser.title": "删除用户",
-  "disabled": "disabled",
+  "disabled": "禁用",
   "expired": "{{time}}失效",
   "expires": "{{time}}有效",
   "header.donation": "捐赠",
@@ -36,7 +36,7 @@
   "header.nodesUsage": "节点统计",
   "hostsDialog.addHost": "添加主机",
   "hostsDialog.advancedOptions": "高级选项",
-  "hostsDialog.allowinsecure": "Allow Insecure",
+  "hostsDialog.allowinsecure": "允许不安全连接",
   "hostsDialog.alpn": "ALPN",
   "hostsDialog.apply": "保存",
   "hostsDialog.currentServer": "当前服务器的 IP 地址",
@@ -48,24 +48,28 @@
   "hostsDialog.fingerprint": "指纹",
   "hostsDialog.fragment": "碎片图案",
   "hostsDialog.fragment.info": "length,interval,packet (e.g. 10-100,100-200,tlshello)",
-  "hostsDialog.fragment.info.attention": "Attention: currently, this feature only supported in streisand >= 1.6.12 and v2rayNG >= 1.8.16 (custom config)",
-  "hostsDialog.fragment.info.examples": "Examples:",
+  "hostsDialog.fragment.info.attention": "请注意： 当前特性仅支持 streisand >= 1.6.12 and v2rayNG >= 1.8.16 (自定义配置)",
+  "hostsDialog.fragment.info.examples": "示例：",
+  "hostsDialog.noise": "噪声模式",
+  "hostsDialog.noise.info": "packet,delay (e.g. rand:10-20,100-200)",
+  "hostsDialog.noise.info.attention": "Attention: currently, this feature only supported in streisand >= 1.6.32 and v2rayNG >= 1.8.39 (custom config)",
+  "hostsDialog.noise.info.examples": "Examples:",
   "hostsDialog.host": "请求主机",
   "hostsDialog.host.info": "默认情况下，如果在 Xray 配置中设置了请求主机，则使用该主机。但是，如果需要，您可以在此处设置自定义请求主机。",
-  "hostsDialog.host.multiHost": "To set multiple addresses, separate them with <badge>,</badge> Each time an address is chosen randomly.",
-  "hostsDialog.host.wildcard": "Use <badge>*</badge> to generate a random string (works for wildcard domain names)",
+  "hostsDialog.host.multiHost": "使用 <badge>,</badge> 作为分隔符设置多个地址，使用时，每次随机选择一个地址。",
+  "hostsDialog.host.wildcard": "使用 <badge>*</badge> 生成随机字符串 (在通配符域名中生效)",
   "hostsDialog.jalaliExpireDate": "用户阳历有效日期",
   "hostsDialog.loading": "加载中...",
   "hostsDialog.muxEnable": "使能够 MUX",
-  "hostsDialog.path": "Path",
-  "hostsDialog.path.info": "Set a path for host users, useful behind a reverse proxy.",
+  "hostsDialog.path": "路径",
+  "hostsDialog.path.info": "为主机用户设置路径，通常在反向代理中使用。",
   "hostsDialog.port": "端口",
   "hostsDialog.port.info": "默认情况下，主机使用入站端口的默认端口。如果此主机是一个服务器，用于从与您的服务器端口不同的端口转发流量，则可以设置自定义端口。例如，服务器可能会将来自端口 8443 的流量转发到您入站服务器的默认端口。",
   "hostsDialog.proxyMethod": "代理传输方法（例如 ws）",
   "hostsDialog.proxyOutbound": "代理出站 json",
   "hostsDialog.proxyOutbound.info": "额外出站（仅在 v2ray 自定义配置中）",
   "hostsDialog.proxyProtocol": "代理协议（例如 VMess）",
-  "hostsDialog.randomUserAgent": "Use random user agent",
+  "hostsDialog.randomUserAgent": "使用随机用户代理",
   "hostsDialog.remainingData": "用户剩余流量情况",
   "hostsDialog.remainingDays": "用户的剩余天数",
   "hostsDialog.remainingTime": "用户剩余时间",
@@ -91,7 +95,7 @@
   "nodes.addNewMarzbanNode": "添加新的 Marzban 节点",
   "nodes.addNode": "添加节点",
   "nodes.addNodeSuccess": "节点 {{name}} 添加成功",
-  "nodes.apply": "editNode",
+  "nodes.apply": "编辑节点",
   "nodes.certificate": "证书",
   "nodes.connection-hint": "要设置Marzban节点，您需要在节点上设置此证书以初始化主服务器和节点之间的安全连接",
   "nodes.download-certificate": "下载证书",
@@ -106,7 +110,7 @@
   "nodes.show-certificate": "如何证书",
   "nodes.title": "使用 Marzban-Node，您可以通过在不同的服务器上添加不同的节点来提供多节点负载。",
   "nodes.usageCoefficient": "使用系数",
-  "on_hold": "On Hold",
+  "on_hold": "在线",
   "password": "密码",
   "previous": "上一页",
   "qrcodeDialog.sublink": "订阅链接",
@@ -120,17 +124,17 @@
   "resetUserUsage.prompt": "您确定要重置 <b>{{username}}</b> 的流量统计吗？",
   "resetUserUsage.success": "{{username}} 的流量统计重置完成。",
   "resetUserUsage.title": "重置用户流量统计",
-  "revoke": "Revoke",
-  "revokeUserSub.error": "Subscription revoke failed, please try again.",
-  "revokeUserSub.prompt": "Are you sure you want to revoke <b>{{username}}</b>'s subscription?",
-  "revokeUserSub.success": "{{username}}'s subscription has revoked successfully.",
-  "revokeUserSub.title": "Revoke User Subscription",
+  "revoke": "撤销",
+  "revokeUserSub.error": "撤销订阅失败，请稍候再试！",
+  "revokeUserSub.prompt": "您确定要撤销用户 <b>{{username}}</b> 的订阅吗？",
+  "revokeUserSub.success": "成功撤销用户 {{username}} 的订阅。",
+  "revokeUserSub.title": "撤销用户订阅",
   "search": "搜索",
-  "status.active": "Active",
-  "status.disabled": "Disabled",
-  "status.expired": "Expired",
-  "status.limited": "Limited",
-  "status.on_hold": "On Hold",
+  "status.active": "活跃",
+  "status.disabled": "禁用",
+  "status.expired": "过期",
+  "status.limited": "受限",
+  "status.on_hold": "在线",
   "nodeModal.status.error": "误",
   "nodeModal.status.disabled": "已禁用",
   "nodeModal.status.connecting": "正在连接",
@@ -147,7 +151,7 @@
   "userDialog.hours": "小时",
   "userDialog.method": "加密方式",
   "userDialog.months": "月",
-  "userDialog.note": "Note",
+  "userDialog.note": "备注",
   "userDialog.onHold": "保持",
   "userDialog.onHoldExpireDuration": "过期时间",
   "userDialog.optional": "可以为空",
@@ -160,7 +164,7 @@
   "userDialog.resetStrategyNo": "无",
   "userDialog.resetStrategyWeekly": "每周",
   "userDialog.resetUsage": "重置流量",
-  "userDialog.revokeSubscription": "Revoke Subscription",
+  "userDialog.revokeSubscription": "撤销订阅",
   "userDialog.selectOneProtocol": "请至少选择一个协议",
   "userDialog.shadowsocksDesc": "快速且安全, 但效率不如其它",
   "userDialog.startDate": "开始日期",
@@ -172,7 +176,7 @@
   "userDialog.userEdited": "已更新用户 {{username}}。",
   "userDialog.vlessDesc": "轻量、快速且安全",
   "userDialog.vmessDesc": "快速且安全",
-  "userDialog.warningNoProtocol": "Please select at least one protocol",
+  "userDialog.warningNoProtocol": "请至少选择一个协议",
   "userDialog.weeks": "周",
   "username": "用户名",
   "users": "用户",
@@ -183,5 +187,6 @@
   "usersTable.noUser": "还没有添加任何用户",
   "usersTable.noUserMatched": "没有找到您搜索的用户",
   "usersTable.status": "状态",
-  "usersTable.total": "总共"
+  "usersTable.sortByExpire": "按过期时间排序",
+  "usersTable.total": "总共",
 }
